requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow>=10.0.0
cryptography>=41.0.0
python-dotenv==1.0.0
Gunicorn==21.2.0
